Move useChat hook to pages
