+++ frontend/src/app/(authorized)/chat/[[...userId]]/useChat.ts
